+++ src/frontend/apps/impress/src/core/config/api/useConfig.tsx
@@ -20,6 +20,7 @@ interface ThemeCustomization {
   header?: HeaderType;
   help: {
     documentation_url?: string;
+    support_mailto?: string;
     legal_links?: {
       personal_data?: string;
       terms_of_use?: string;
@@ -52,6 +53,7 @@ export interface ConfigResponse {
   CONVERSION_FILE_EXTENSIONS_ALLOWED: string[];
   CONVERSION_FILE_MAX_SIZE: number;
   CONVERSION_UPLOAD_ENABLED?: boolean;
+  CRISP_WEBSITE_ID?: string;
   ENVIRONMENT: string;
   FRONTEND_CSS_URL?: string;
   FRONTEND_HOMEPAGE_FEATURE_ENABLED?: boolean;